feat(m9.4): ABAC v1 — OPA/WASM policy engine + 5 attribute-aware policies
Adds Rego-based attribute authorization coexisting with the flat RBAC
layer from M9.3. Five permissions get attribute conditions on top of the
flat-permission check; the other 10 stay flat booleans.

Bundle:
- policies/stadium/authz.rego — Rego v1 entrypoint stadium/authz/allow
- policies/stadium/authz_test.rego — 19 unit tests (all pass)
- scripts/install-opa.sh — pins OPA v1.18.2 (darwin/linux, arm64/amd64)
- scripts/build-policies.ts — opa build -t wasm + extract policy.wasm

Runtime (netlify/lib/policy.ts):
- Lazy-loaded @open-policy-agent/opa-wasm evaluator, cached per warm
  function instance
- decide(action, subject, resource): boolean — fail-closed on errors
- subjectFromClaims() helper for assembling PolicySubject from JWT
- isAbacScoped() gate so non-ABAC actions skip the OPA call entirely

Wired into netlify/functions/mutations.ts:
- incident:transition — manager tier>=4 only; admin/superadmin any tier
- dispatch:create — manager same-zone only; admin/superadmin any target
- dispatch:update — self (target_staff_phone match) OR admin/manager/superadmin
- staff:reassign — manager source+target zone match; admin/superadmin any

Schema fix:
- audit_ledger.timestamp changed integer → bigint (was overflowing at
  ms-since-epoch values; original SQL schema was BIGINT, Drizzle
  regen had downgraded it to integer)
- Fresh migration 0000_perpetual_kid_colt.sql replaces 0000_minor_komodo.sql

Matrix update:
- manager role granted incident:transition at flat level so the ABAC
  tier-gate can restrict to tier 4-5 (previously the flat check rejected
  managers before ABAC ran)

Tests: 19/19 Rego + 82/82 TS + tsc clean.

Note: vite.config.ts modified locally (mkcert + stadops.local) is
intentionally NOT committed — it's a per-developer local preference.
Add '127.0.0.1 stadops.local' to /etc/hosts and run 'mkcert -install'
to use the HTTPS local dev setup.

+++ database/migrations/0000_perpetual_kid_colt.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "audit_ledger" (
 	"event_id" text PRIMARY KEY NOT NULL,
 	"tenant_id" text NOT NULL,
-	"timestamp" integer NOT NULL,
+	"timestamp" bigint NOT NULL,
 	"actor_uid" text NOT NULL,
 	"actor_role" text NOT NULL,
 	"actor_phone_email" text,
@@ -160,13 +160,9 @@ CREATE INDEX "idx_staff_tenant" ON "staff_roster" USING btree ("tenant_id","stat
 CREATE INDEX "idx_memberships_tenant" ON "tenant_memberships" USING btree ("tenant_id");--> statement-breakpoint
 CREATE INDEX "idx_user_perms_user" ON "user_permissions" USING btree ("user_id");--> statement-breakpoint
 CREATE INDEX "idx_users_global_role" ON "users" USING btree ("global_role");--> statement-breakpoint
-CREATE INDEX "idx_users_status" ON "users" USING btree ("status");
---> statement-breakpoint
+CREATE INDEX "idx_users_status" ON "users" USING btree ("status");--> statement-breakpoint
 
 -- ─── Triggers (not expressible in Drizzle schema definitions) ────────────────
--- Carried forward from legacy 0001_init.sql + 0003_audit_ledger.sql.
-
--- updated_at auto-touch function (used by incidents, staff_roster, users).
 CREATE OR REPLACE FUNCTION touch_updated_at()
 RETURNS TRIGGER AS $$
 BEGIN
@@ -187,7 +183,6 @@ DROP TRIGGER IF EXISTS trg_users_touch ON users;
 CREATE TRIGGER trg_users_touch BEFORE UPDATE ON users
     FOR EACH ROW EXECUTE FUNCTION touch_updated_at();--> statement-breakpoint
 
--- Audit ledger WORM enforcement (ADR-0005, ADR-0009) — rejects UPDATE/DELETE.
 CREATE OR REPLACE FUNCTION reject_audit_mutation()
 RETURNS TRIGGER AS $$
 BEGIN
